move `replace_lines` from `snapbox` to `trycmd`
