+++ crates/snapbox/src/data.rs
@@ -65,36 +65,6 @@ impl Data {
             .map_err(|e| format!("Failed to write {}: {}", path.display(), e).into())
     }
 
-    /// Update an inline snapshot
-    pub fn replace_lines(
-        &mut self,
-        line_nums: std::ops::Range<usize>,
-        text: &str,
-    ) -> Result<(), crate::Error> {
-        let mut output_lines = String::new();
-
-        let s = self
-            .as_str()
-            .ok_or("Binary file can't have lines replaced")?;
-        for (line_num, line) in crate::utils::LinesWithTerminator::new(s)
-            .enumerate()
-            .map(|(i, l)| (i + 1, l))
-        {
-            if line_num == line_nums.start {
-                output_lines.push_str(text);
-                if !text.is_empty() && !text.ends_with('\n') {
-                    output_lines.push('\n');
-                }
-            }
-            if !line_nums.contains(&line_num) {
-                output_lines.push_str(line);
-            }
-        }
-
-        *self = Self::text(output_lines);
-        Ok(())
-    }
-
     /// Post-process text
     ///
     /// See [utils][crate::utils]
@@ -238,69 +208,4 @@ fn is_binary(data: &[u8]) -> bool {
 #[cfg(not(feature = "detect-encoding"))]
 fn is_binary(_data: &[u8]) -> bool {
     false
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn replace_lines_same_line_count() {
-        let input = "One\nTwo\nThree";
-        let line_nums = 2..3;
-        let replacement = "World\n";
-        let expected = Data::text("One\nWorld\nThree");
-
-        let mut actual = Data::text(input);
-        actual.replace_lines(line_nums, replacement).unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn replace_lines_grow() {
-        let input = "One\nTwo\nThree";
-        let line_nums = 2..3;
-        let replacement = "World\nTrees\n";
-        let expected = Data::text("One\nWorld\nTrees\nThree");
-
-        let mut actual = Data::text(input);
-        actual.replace_lines(line_nums, replacement).unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn replace_lines_shrink() {
-        let input = "One\nTwo\nThree";
-        let line_nums = 2..3;
-        let replacement = "";
-        let expected = Data::text("One\nThree");
-
-        let mut actual = Data::text(input);
-        actual.replace_lines(line_nums, replacement).unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn replace_lines_no_trailing() {
-        let input = "One\nTwo\nThree";
-        let line_nums = 2..3;
-        let replacement = "World";
-        let expected = Data::text("One\nWorld\nThree");
-
-        let mut actual = Data::text(input);
-        actual.replace_lines(line_nums, replacement).unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn replace_lines_empty_range() {
-        let input = "One\nTwo\nThree";
-        let line_nums = 2..2;
-        let replacement = "World\n";
-        let expected = Data::text("One\nWorld\nTwo\nThree");
-
-        let mut actual = Data::text(input);
-        actual.replace_lines(line_nums, replacement).unwrap();
-        assert_eq!(expected, actual);
-    }
 }
